comments and replies now notify user/seller via email

+++ paw-webapp/persistence/src/main/java/ar/edu/itba/paw/persistence/ProductHibernateDao.java
@@ -98,7 +98,7 @@ public class ProductHibernateDao implements ProductDao {
         for(Object o : jpaQuery.getResultList()) {
 //            BigInteger big = BigInteger.valueOf((Integer)o);
 //            products.add(big.longValue());
-            products.add(((BigInteger)o).longValue());
+            products.add(((Integer)o).longValue());
         }
